Move Section model to the physionet app.
