+++ physionet-django/console/static/console/css/console.css
@@ -543,7 +543,3 @@ footer.sticky-footer {
 .table-process {
   width: 11%;
 }
-
-.django-ckeditor-widget{
-	width: 100%;
-}
