+++ physionet-django/project/models.py
@@ -9,4 +9,3 @@ from project.modelcomponents.metadata import *
 from project.modelcomponents.publishedproject import *
 from project.modelcomponents.storage import *
 from project.modelcomponents.submission import *
-from project.modelcomponents.templates import *
